+++ src/components/Select/types.ts
@@ -82,7 +82,16 @@ export type SelectProps<T = any> = QAProps &
         value?: string[];
         defaultValue?: string[];
         options?: (SelectOption<T> | SelectOptionGroup<T>)[];
+        /**
+         * @deprecated Prop `error` has a lower priority than `errorMessage`. Use `errorMessage` instead
+         */
         error?: string | boolean;
+        /** Determines content of the error message */
+        errorMessage?: React.ReactNode;
+        /** Determines whether the error message will be placed under the input field as text or in the tooltip */
+        errorPlacement?: 'outside' | 'inside';
+        /** Describes the validation state */
+        validationState?: 'invalid';
         multiple?: boolean;
         filterable?: boolean;
         disablePortal?: boolean;